fix(moments): one estimator per file
