+++ src/skfolio/moments/covariance/__init__.py
@@ -3,18 +3,16 @@
 from skfolio.moments.covariance._base import (
     BaseCovariance,
 )
-from skfolio.moments.covariance._covariance import (
-    OAS,
-    DenoiseCovariance,
-    DetoneCovariance,
-    EWCovariance,
-    EmpiricalCovariance,
-    GerberCovariance,
-    GraphicalLassoCV,
-    ImpliedCovariance,
-    LedoitWolf,
-    ShrunkCovariance,
-)
+from skfolio.moments.covariance._denoise_covariance import DenoiseCovariance
+from skfolio.moments.covariance._detone_covariance import DetoneCovariance
+from skfolio.moments.covariance._empirical_covariance import EmpiricalCovariance
+from skfolio.moments.covariance._ew_covariance import EWCovariance
+from skfolio.moments.covariance._gerber_covariance import GerberCovariance
+from skfolio.moments.covariance._graphical_lasso_cv import GraphicalLassoCV
+from skfolio.moments.covariance._implied_covariance import ImpliedCovariance
+from skfolio.moments.covariance._ledoit_wolf import LedoitWolf
+from skfolio.moments.covariance._oas import OAS
+from skfolio.moments.covariance._shrunk_covariance import ShrunkCovariance
 
 __all__ = [
     "BaseCovariance",
