+++ src/skfolio/moments/expected_returns/__init__.py
@@ -3,13 +3,10 @@
 from skfolio.moments.expected_returns._base import (
     BaseMu,
 )
-from skfolio.moments.expected_returns._expected_returns import (
-    EWMu,
-    EmpiricalMu,
-    EquilibriumMu,
-    ShrunkMu,
-    ShrunkMuMethods,
-)
+from skfolio.moments.expected_returns._empirical_mu import EmpiricalMu
+from skfolio.moments.expected_returns._equilibrium_mu import EquilibriumMu
+from skfolio.moments.expected_returns._ew_mu import EWMu
+from skfolio.moments.expected_returns._shrunk_mu import ShrunkMu, ShrunkMuMethods
 
 __all__ = [
     "BaseMu",
