+++ src/skfolio/moments/expected_returns/_shrunk_mu.py
@@ -1,4 +1,4 @@
-"""Expected returns estimators."""
+"""Shrinkage Expected Returns (Mu) Estimators."""
 
 # Copyright (c) 2023
 # Author: Hugo Delatte <delatte.hugo@gmail.com>
@@ -12,207 +12,10 @@ from enum import auto
 
 import numpy as np
 import numpy.typing as npt
-import pandas as pd
 
 from skfolio.moments.covariance import BaseCovariance, EmpiricalCovariance
 from skfolio.moments.expected_returns._base import BaseMu
 from skfolio.utils.tools import AutoEnum, check_estimator
-
-
-class EmpiricalMu(BaseMu):
-    """Empirical Expected Returns (Mu) estimator.
-
-    Estimates the expected returns with the historical mean.
-
-    Parameters
-    ----------
-    window_size : int, optional
-        Window size. The model is fitted on the last `window_size` observations.
-        The default (`None`) is to use all the data.
-
-    Attributes
-    ----------
-    mu_ : ndarray of shape (n_assets,)
-        Estimated expected returns of the assets.
-
-    n_features_in_ : int
-        Number of assets seen during `fit`.
-
-    feature_names_in_ : ndarray of shape (`n_features_in_`,)
-        Names of assets seen during `fit`. Defined only when `X`
-        has assets names that are all strings.
-    """
-
-    def __init__(self, window_size: int | None = None):
-        self.window_size = window_size
-
-    def fit(self, X: npt.ArrayLike, y=None, **fit_params) -> "EmpiricalMu":
-        """Fit the Mu Empirical estimator model.
-
-        Parameters
-        ----------
-        X : array-like of shape (n_observations, n_assets)
-            Price returns of the assets.
-
-        y : Ignored
-            Not used, present for API consistency by convention.
-
-        Returns
-        -------
-        self : EmpiricalMu
-            Fitted estimator.
-        """
-        X = self._validate_data(X)
-        if self.window_size is not None:
-            X = X[-self.window_size :]
-        self.mu_ = np.mean(X, axis=0)
-        return self
-
-
-class EWMu(BaseMu):
-    r"""Exponentially Weighted Expected Returns (Mu) estimator.
-
-    Estimates the expected returns with the exponentially weighted mean (EWM).
-
-    Parameters
-    ----------
-    window_size : int, optional
-        Window size. The model is fitted on the last `window_size` observations.
-        The default (`None`) is to use all the data.
-
-    alpha : float, default=0.2
-        Exponential smoothing factor. The default value is `0.2`.
-
-        :math:`0 < \alpha \leq 1`.
-
-    Attributes
-    ----------
-    mu_ : ndarray of shape (n_assets,)
-       Estimated expected returns of the assets.
-
-    n_features_in_ : int
-        Number of assets seen during `fit`.
-
-    feature_names_in_ : ndarray of shape (`n_features_in_`,)
-        Names of assets seen during `fit`. Defined only when `X`
-        has assets names that are all strings.
-    """
-
-    def __init__(self, window_size: int | None = None, alpha: float = 0.2):
-        self.window_size = window_size
-        self.alpha = alpha
-
-    def fit(self, X: npt.ArrayLike, y=None, **fit_params) -> "EWMu":
-        """Fit the EWMu estimator model.
-
-        Parameters
-        ----------
-        X : array-like of shape (n_observations, n_assets)
-           Price returns of the assets.
-
-        y : Ignored
-            Not used, present for API consistency by convention.
-
-        Returns
-        -------
-        self : EWMu
-            Fitted estimator.
-        """
-        X = self._validate_data(X)
-        if self.window_size is not None:
-            X = X[-self.window_size :]
-        self.mu_ = pd.DataFrame(X).ewm(alpha=self.alpha).mean().iloc[-1, :].to_numpy()
-        return self
-
-
-class EquilibriumMu(BaseMu):
-    r"""Equilibrium Expected Returns (Mu) estimator.
-
-    The Equilibrium is defined as:
-
-        .. math:: risk\_aversion \times \Sigma \cdot w^T
-
-    For Market Cap Equilibrium, the weights are the assets Market Caps.
-    For Equal-weighted Equilibrium, the weights are equal-weighted (1/N).
-
-    Parameters
-    ----------
-    risk_aversion : float, default=1.0
-        Risk aversion factor.
-        The default value is `1.0`.
-
-    weights : array-like of shape (n_assets,), optional
-        Asset weights used to compute the Expected Return Equilibrium.
-        The default is to use the equal-weighted equilibrium (1/N).
-        For a Market Cap weighted equilibrium, you must provide the asset Market Caps.
-
-    covariance_estimator : BaseCovariance, optional
-        :ref:`Covariance estimator <covariance_estimator>` used to estimate the
-        covariance in the equilibrium formula.
-        The default (`None`) is to use :class:`~skfolio.moments.EmpiricalCovariance`.
-
-    Attributes
-    ----------
-    mu_ : ndarray of shape (n_assets,)
-          Estimated expected returns of the assets.
-
-    covariance_estimator_ : BaseCovariance
-        Fitted `covariance_estimator`.
-
-    n_features_in_ : int
-        Number of assets seen during `fit`.
-
-    feature_names_in_ : ndarray of shape (`n_features_in_`,)
-        Names of assets seen during `fit`. Defined only when `X`
-        has assets names that are all strings.
-    """
-
-    covariance_estimator_: BaseCovariance
-
-    def __init__(
-        self,
-        risk_aversion: float = 1,
-        weights: np.ndarray | None = None,
-        covariance_estimator: BaseCovariance | None = None,
-    ):
-        self.risk_aversion = risk_aversion
-        self.weights = weights
-        self.covariance_estimator = covariance_estimator
-
-    def fit(self, X: npt.ArrayLike, y=None, **fit_params) -> "EquilibriumMu":
-        """Fit the EquilibriumMu estimator model.
-
-        Parameters
-        ----------
-        X : array-like of shape (n_observations, n_assets)
-           Price returns of the assets.
-
-        y : Ignored
-            Not used, present for API consistency by convention.
-
-        Returns
-        -------
-        self : EquilibriumMu
-            Fitted estimator.
-        """
-        # fitting estimators
-        self.covariance_estimator_ = check_estimator(
-            self.covariance_estimator,
-            default=EmpiricalCovariance(),
-            check_type=BaseCovariance,
-        )
-        self.covariance_estimator_.fit(X)
-
-        # we validate and convert to numpy after all models have been fitted to keep
-        # features names information.
-        X = self._validate_data(X)
-        n_assets = X.shape[1]
-        if self.weights is None:
-            weights = np.ones(n_assets) / n_assets
-        else:
-            weights = np.asarray(self.weights)
-        self.mu_ = self.risk_aversion * self.covariance_estimator_.covariance_ @ weights
-        return self
 
 
 class ShrunkMuMethods(AutoEnum):
